% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARGOS2.R
\name{ARGOS2}
\alias{ARGOS2}
\alias{ARGOS2.downloadNew}
\alias{ARGOS2.update_incoming}
\title{ARGOS2}
\usage{
ARGOS2.downloadNew()

ARGOS2.update_incoming(x)
}
\arguments{
\item{X}{data.table creaded by ARGOS2.downloadNew}
}
\description{
ARGOS2
}
