% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_BTatWESTERHOLZ.R
\name{BT_at_WESTERHOLZ_change_ID}
\alias{BT_at_WESTERHOLZ_change_ID}
\title{uses ID_changes table
Run this function after each update in ID_changes (run on cron!)}
\usage{
BT_at_WESTERHOLZ_change_ID(h = getOption("host"))
}
\description{
uses ID_changes table
Run this function after each update in ID_changes (run on cron!)
}
