% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_DBIO.R
\name{db_copy}
\alias{db_copy}
\title{db_copy
database copy from src to dst. It has to be run on dst.}
\usage{
db_copy(db, src, dst, cnf = config::get())
}
\arguments{
\item{src}{source host}

\item{dst}{destination host}

\item{cnf}{configuration variables are
obtained from an external config file.
default to config::get().}
}
\description{
db_copy
database copy from src to dst. It has to be run on dst.
}
\examples{
\dontrun{
require(dup)
Sys.setenv(R_CONFIG_ACTIVE = "localhost")
db_copy(db = "FIELD_2024_NOLAatDUMMERSEE", src = "remotehost_1", dst ="host")
}

}
