% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRUID.R
\name{DRUID.downloadNew}
\alias{DRUID.downloadNew}
\title{Download the latest DRUID data
Fetch DRUID data}
\usage{
DRUID.downloadNew(what)
}
\arguments{
\item{what}{gps, odba or env}
}
\description{
Download the latest DRUID data
Fetch DRUID data
}
\examples{
x = DRUID.downloadNew()
}
