% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_PIPELINES.R
\name{dup}
\alias{dup}
\alias{ARGOS.pipeline}
\alias{DB_internal_updates.pipeline}
\alias{backup.pipeline}
\alias{RUFFatSEEWIESEN_photos.pipeline}
\alias{Argos2.pipeline}
\title{Database Unattended Pipelines}
\usage{
ARGOS.pipeline()

DB_internal_updates.pipeline()

backup.pipeline(cnf = config::get("host"))

RUFFatSEEWIESEN_photos.pipeline(...)

Argos2.pipeline()
}
\description{
Database Unattended Pipelines

Argos pipeline

DB internal updates pipeline

backup pipeline

RUFFatSEEWIESEN pipelines

DRUID pipeline
}
\note{
On error all pipelines should return FALSE and send a pushover
}
