% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_pipelines.R
\name{export_to_mapping.pipeline}
\alias{export_to_mapping.pipeline}
\title{Export Argos for mapping}
\usage{
export_to_mapping.pipeline(
  db_tabs = c("2019_LBDO", "2020_BADO", "2022_WRSA"),
  ...
)
}
\description{
Export Argos for mapping
}
