% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_email.R
\name{extract_email_attachements}
\alias{extract_email_attachements}
\title{extract_email_attachements to their dirs AFTER calling offlineimap}
\usage{
extract_email_attachements(
  maildir,
  onlynew = TRUE,
  keep = ".*\\\\.(TXT|txt|csv|CSV)",
  cnf = config::get()
)
}
\arguments{
\item{maildir}{ARGOS or GSM_MTI}

\item{onlynew}{default to TRUE}

\item{keep}{default to ".*\\.(TXT|txt)"}

\item{cnf}{configuration variables are obtained from an external file config file.
default to config::get().}
}
\description{
extract_email_attachements to their dirs AFTER calling offlineimap
}
\note{
When offlineimap() is called independently then, call extract_email_attachements(onlynew = FALSE) once so that
no new emails are disregarded.
}
\examples{
\dontrun{
 dup::extract_email_attachements(maildir = 'ARGOS', onlynew = FALSE)
 extract_email_attachements(maildir = 'GSM_MTI', onlynew = FALSE)

}
}
