% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_DBIO.R
\name{mariacon}
\alias{mariacon}
\title{Makes a database connection based on a config file}
\usage{
mariacon(db)
}
\value{
a MariaDB connection object
}
\description{
Makes a database connection based on a config file
}
