% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_GPS.R
\name{mti_gps.BUTEOatEUROPE}
\alias{mti_gps.BUTEOatEUROPE}
\alias{scidbupdate_mti_gps.BUTEOatEUROPE}
\title{mti_gps.BUTEOatEUROPE}
\usage{
scidbupdate_mti_gps.BUTEOatEUROPE(host = getOption("host"),
  user = "mihai")
}
\description{
mti_gps.BUTEOatEUROPE
}
\examples{
scidbupdate_mti_gps.BUTEOatEUROPE()
}
