% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_GPS.R
\name{mti_sensors.BUTEOatEUROPE}
\alias{mti_sensors.BUTEOatEUROPE}
\alias{scidbupdate_mti_sensors.BUTEOatEUROPE}
\title{mti_sensors.BUTEOatEUROPE}
\usage{
scidbupdate_mti_sensors.BUTEOatEUROPE(host = getOption("host"),
  user = "mihai")
}
\description{
mti_sensors.BUTEOatEUROPE
}
\examples{
scidbupdate_mti_sensors.BUTEOatEUROPE()
}
