% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_DBIO.R
\name{mysqldump}
\alias{mysqldump}
\title{mysqldump}
\usage{
mysqldump(
  db,
  tables,
  user,
  pwd,
  host = "127.0.0.1",
  filenam = "dbdump.sql",
  dir = getwd(),
  dryrun = FALSE,
  compress = TRUE,
  ...
)
}
\arguments{
\item{db}{db}

\item{tables}{tables are given as a "tableName1 tableName2".}

\item{user}{user}

\item{pwd}{pwd}

\item{host}{default to '127.0.0.1'}

\item{filenam}{filenam.}

\item{dir}{saving location on disk.}

\item{dryrun}{when TRUE return call only. default to FALSE.}

\item{compress}{when TRUE archive the sql output. default to TRUE.}

\item{...}{further arguments to mysqldump
(e.g. --no-data --no-create-db)}
}
\value{
the file path to the sql file or system call
when dryrun = TRUE
}
\description{
mysqldump
}
\examples{
\dontrun{
fp = mysqldump('tests',  user = 'testuser', pwd = '', dir = tempdir(), filenam = "tempbk.sql" , dryrun = TRUE)
mysqldump('tests', 't1 t2', 'testuser', pwd = '', dir = tempdir(), filenam = "tempbk.sql" , dryrun = TRUE)
mysqldump('tests', 't1', 'testuser', pwd = '', dir = tempdir(), filenam = "tempbk.sql", compress = FALSE , dryrun = TRUE)
}


}
