% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_backup_sqldump.R
\name{mysqldump_host}
\alias{mysqldump_host}
\title{mysqldump_host}
\usage{
mysqldump_host(cnf = config::get(), exclude = c("mysql",
  "information_schema", "performance_schema", "phpmyadmin"),
  parallel = TRUE)
}
\arguments{
\item{cnf}{configuration variables are obtained from an external file config file. 
default to config::get().}

\item{exclude}{db-s to exclude default to c('mysql', 'information_schema', 'performance_schema')}

\item{parallel}{default to TRUE}
}
\description{
mysqldump_host
}
\examples{
\dontrun{
Sys.setenv(R_CONFIG_ACTIVE = "default")
dup::mysqldump_host()

}
}
