% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_DBIO.R
\name{mysqlrestore}
\alias{mysqlrestore}
\title{mysqlrestore}
\usage{
mysqlrestore(file, db, user, pwd, host = "127.0.0.1", dryrun = FALSE)
}
\arguments{
\item{file}{sql or sql.gz file}

\item{db}{database name}

\item{user}{user, default to 'root'}

\item{host}{default to  '127.0.0.1'}

\item{dryrun}{only print the mariadb cli call and exit}
}
\description{
restore sql file locally
}
