% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_backup_sqldump.R
\name{mysqlrestore_host}
\alias{mysqlrestore_host}
\title{mysqlrestore_host}
\usage{
mysqlrestore_host(
  cnf = config::get(),
  backup,
  wipe = FALSE,
  restore_users = FALSE,
  parallel = TRUE,
  exclude,
  host_is_set = FALSE
)
}
\arguments{
\item{cnf}{configuration variables are obtained from an external file config file. 
default to config::get().}

\item{backup}{path to backup dir. if missing the last backup is used.}

\item{wipe}{drop all non-system db-s before restore. default to FALSE}

\item{restore_users}{restore mysql.user table}

\item{parallel}{default to TRUE}

\item{exclude}{db to exclude from restoring}

\item{host_is_set}{default to FALSE}

\item{...}{further options passed to mysqlrestore}
}
\description{
restore an entire db system or several db-s
}
\examples{
\dontrun{

 require(dup)
 Sys.setenv(R_CONFIG_ACTIVE = "localhost")
 mysqlrestore_host()
}


}
