% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{read_boxtxt}
\alias{read_boxtxt}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{    read SNB data
}\if{html}{\out{</div>}}}
\usage{
read_boxtxt(f)
}
\arguments{
\item{f}{path to the snb file}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a  \code{data.table} .
}\if{html}{\out{</div>}}
}
\description{
read an SNB file as data.table-s (only transponder and Light barrier data are parsed).
}
\examples{
x = read_boxtxt(system.file('test_files_SNB', '80', 'BOX0080.TXT', package = 'SNB2'))
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   MV
}\if{html}{\out{</div>}}
}
