% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_through_email.R
\name{read_email_attachements}
\alias{read_email_attachements}
\title{read_attachements}
\usage{
read_email_attachements(maildir, exclude, pattern, lastdate,
  sepDate = "", ...)
}
\arguments{
\item{maildir}{maildir}

\item{exclude}{exclude filenames (if they are in the db already)}

\item{pattern}{regexp on file name.}

\item{lastdate}{date of the last included file (date is embedded in the filename)}

\item{...}{passed to fread}

\item{cnf}{configuration variables are obtained from an external file config file. 
default to config::get().}
}
\value{
DT
}
\description{
read_attachements
}
\examples{
x = read_email_attachements(maildir='ARGOS'  , sep = ";")
x = read_email_attachements(maildir='GSM_MTI', pattern = "g_")
x = read_email_attachements(maildir='GSM_MTI', pattern = "e_")
}
