% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_DBIO.R
\name{rm_old_backups}
\alias{rm_old_backups}
\title{rm_old_backups}
\usage{
rm_old_backups(path = config::get("dir")$backupdir, keep = 10)
}
\arguments{
\item{path}{path to backup dir. default taken from the config.yml file retrieved by}

\item{keep}{how many prior backups to keep}
}
\value{
names of removed backups
}
\description{
remove ond backups
}
