% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_internal_updates.R
\name{RUFF_at_SEEWIESEN_change_ID}
\alias{RUFF_at_SEEWIESEN_change_ID}
\title{uses ID_changes table}
\usage{
RUFF_at_SEEWIESEN_change_ID(cnf = config::get())
}
\arguments{
\item{cnf}{configuration variables are obtained from an external file config file.
default to config::get().}
}
\value{
Nrows updated in DB_changes
}
\description{
uses ID_changes table
}
