% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUFFatSEEWIESEN.R
\name{RUFFatSEEWIESEN.change_ID}
\alias{RUFFatSEEWIESEN.change_ID}
\title{uses ID_changes table}
\usage{
RUFFatSEEWIESEN.change_ID(cnf = config::get())
}
\arguments{
\item{cnf}{configuration variables are obtained from an external file config file.
default to config::get().}
}
\description{
uses ID_changes table
}
