% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUFFatSEEWIESEN.R
\name{RUFFatSEEWIESEN.photos_prepare}
\alias{RUFFatSEEWIESEN.photos_prepare}
\title{Prepare data for RUFFatSEEWIESEN.photos
Expands ADULTS table by photo ID}
\usage{
RUFFatSEEWIESEN.photos_prepare(
  last_pk,
  basepath = config::get()$dir$ruff_photos
)
}
\description{
Prepare data for RUFFatSEEWIESEN.photos
Expands ADULTS table by photo ID
}
\examples{
x = RUFFatSEEWIESEN.photos_prepare()
}
