% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUFFatSEEWIESEN.R
\name{RUFFatSEEWIESEN.photos_update}
\alias{RUFFatSEEWIESEN.photos_update}
\title{update RUFFatSEEWIESEN.photos
expands ADULTS table by photo ID}
\usage{
RUFFatSEEWIESEN.photos_update()
}
\value{
number of rows updated
}
\description{
update RUFFatSEEWIESEN.photos
expands ADULTS table by photo ID
}
\examples{
x = RUFFatSEEWIESEN.photos_update()
}
