% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_pipelines.R
\name{RUFFatSEEWIESEN_photos.pipeline}
\alias{RUFFatSEEWIESEN_photos.pipeline}
\title{title ruff photos pipeline}
\usage{
RUFFatSEEWIESEN_photos.pipeline(db = "RUFFatSEEWIESEN_photos.sqlite")
}
\description{
title ruff photos pipeline
}
