% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rw2webp}
\alias{rw2webp}
\title{Convert RW2 photos to WebP}
\usage{
rw2webp(src, dest, resizefact = 2, cropamount = 5)
}
\arguments{
\item{src}{output file name}

\item{dest}{output file name}

\item{resizefact}{resize factor (default to 2, half the size of the original image)}

\item{cropamount}{crop amount}

\item{width}{output width in px default to 2000}
}
\description{
Convert RW2 photos to WebP
}
