% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNB.R
\name{scidb_snbUpdater.transponders}
\alias{scidb_snbUpdater.transponders}
\title{b000 tables to transponders table pipeline}
\usage{
scidb_snbUpdater.transponders(cnf = config::get())
}
\arguments{
\item{cnf}{configuration variables are obtained from an external file config file.
default to config::get().}
}
\value{
TRUE on success
}
\description{
b000 tables to transponders table pipeline
}
\examples{
\dontrun{
 require(dup)
 
 scidb_snbUpdater.transponders()
}

}
