% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_email_ARGOS.R
\name{scidbupdate_ARGOS.flush_incoming}
\alias{scidbupdate_ARGOS.flush_incoming}
\title{Argos: move from incoming to YYYY_SSSS}
\usage{
scidbupdate_ARGOS.flush_incoming(cnf = config::get())
}
\arguments{
\item{cnf}{configuration variables are obtained from an external file config file. 
default to config::get().}
}
\description{
Argos: move from incoming to YYYY_SSSS
}
\examples{
scidbupdate_ARGOS.flush_incoming()
}
