% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_email_ARGOS.R
\name{scidbupdate_ARGOS.incoming}
\alias{scidbupdate_ARGOS.incoming}
\title{Argos incoming}
\usage{
scidbupdate_ARGOS.incoming(cnf = config::get(), daysBefore = 365)
}
\arguments{
\item{cnf}{configuration variables are obtained from an external file config file. 
default to config::get().}

\item{daysBefore}{converted to  Sys.Date() - 365 and passed to read_email_attachements(lastdate=).}

\item{user}{default mihai}
}
\description{
Argos incoming
}
\examples{
scidbupdate_ARGOS.incoming()
}
