% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_email_GPS.R
\name{scidbupdate_mti_gps.BUTEOatEUROPE}
\alias{scidbupdate_mti_gps.BUTEOatEUROPE}
\title{mti_gps.BUTEOatEUROPE}
\usage{
scidbupdate_mti_gps.BUTEOatEUROPE(cnf = config::get())
}
\arguments{
\item{cnf}{configuration variables are obtained from an external file config file. 
default to config::get().}
}
\description{
mti_gps.BUTEOatEUROPE
}
\examples{
scidbupdate_mti_gps.BUTEOatEUROPE()
}
