% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_archive.R
\name{scidbupdate_mti_sensors.BUTEOatEUROPE}
\alias{scidbupdate_mti_sensors.BUTEOatEUROPE}
\title{mti_sensors.BUTEOatEUROPE}
\usage{
scidbupdate_mti_sensors.BUTEOatEUROPE(cnf = config::get())
}
\arguments{
\item{cnf}{configuration variables are obtained from an external file config file.
default to config::get().}
}
\description{
mti_sensors.BUTEOatEUROPE
}
\examples{
scidbupdate_mti_sensors.BUTEOatEUROPE()
}
