% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_backup_sqlite.R
\name{sqlitedump}
\alias{sqlitedump}
\title{sqlitedump}
\usage{
sqlitedump(
  db,
  tables,
  exclude_columns,
  indices,
  dir = ".",
  filename = glue("{dir}/{db}.sqlite"),
  remote = TRUE,
  cnf = config::get()
)
}
\arguments{
\item{db}{db name}

\item{tables}{one of more table names, character vector}

\item{exclude_columns}{which columns to exclude}

\item{indices}{which columns are indices}

\item{dir}{directory path. the destination file (same location on both local and remote).}

\item{filename}{name of the sqlite file.}

\item{remote}{when TRUE the file is uploaded to a remote host defined in cnf}

\item{cnf}{configuration variables (host, user, pwd, remotehost) are obtained 
from an external file config file. default to config::get().}
}
\value{
path of the dumped file
}
\description{
sqlitedump
}
\examples{
\dontrun{

	require(dup)
 Sys.setenv(R_CONFIG_ACTIVE = "localhost")
sqlitedump(db = 'ARGOS',
tables = c('2019_LBDO', '2020_BADO', '2022_WRSA'), 
exclude_columns = c("satellite",  "messageDate", "locationClass", "compressionIndex",  "S1", "S2", "S3", "S4", "S5", #' "S6", "S7", "S8", "filenam"), 
indices = c('tagID', 'locationDate')
)
}

}
