% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_backup_txt.R
\name{txtdump}
\alias{txtdump}
\title{txtdump}
\usage{
txtdump(db, table, dir = "~", remote_host, cnf = config::get())
}
\arguments{
\item{db}{db name}

\item{table}{table name}

\item{dir}{directory path}

\item{remote_host}{when given will ssh table to dir on remote_host}

\item{cnf}{configuration variables are obtained from an external file config file. 
default to config::get().}
}
\value{
TRUE on success FALSE on failure
}
\description{
txtdump
}
\examples{
\dontrun{

	require(dup)
 Sys.setenv(R_CONFIG_ACTIVE = "localhost")
 txtdump(db = 'ARGOS', table = '2019_LBDO')
}

}
