% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_internal_updates.R
\name{BT_at_WESTERHOLZ_change_ID}
\alias{BT_at_WESTERHOLZ_change_ID}
\title{uses ID_changes table}
\usage{
BT_at_WESTERHOLZ_change_ID(cnf = config::get())
}
\arguments{
\item{cnf}{configuration variables are obtained from an external file config file.
default to config::get().}
}
\description{
uses ID_changes table
}
