% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_DBIO.R
\name{db_copy}
\alias{db_copy}
\title{db_copy
copy from one host to another. should be run from 'dst'.}
\usage{
db_copy(db, src, dst, cnf = config::get())
}
\arguments{
\item{src}{source host}

\item{dst}{destination host}

\item{cnf}{configuration variables are
obtained from an external config file.
default to config::get().}
}
\description{
db_copy
copy from one host to another. should be run from 'dst'.
}
\examples{
\dontrun{
require(dup)
Sys.setenv(R_CONFIG_ACTIVE = "localhost")
db_copy("FIELD_BTatWESTERHOLZ", "remotehost_1", "host")
}

}
