% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_through_email.R
\name{extract_email_attachements}
\alias{extract_email_attachements}
\title{extract_email_attachements to their dirs AFTER calling offlineimap}
\usage{
extract_email_attachements(maildir, onlynew = TRUE,
  keep = ".*\\\\.(TXT|txt|csv|CSV)", cnf = config::get())
}
\arguments{
\item{maildir}{ARGOS or GSM_MTI}

\item{onlynew}{default to TRUE}

\item{keep}{default to ".*\\.(TXT|txt)"}

\item{cnf}{configuration variables are obtained from an external file config file. 
default to config::get().}
}
\description{
extract_email_attachements to their dirs AFTER calling offlineimap
}
\examples{
\dontrun{
 extract_email_attachements(maildir = 'ARGOS', onlynew = TRUE)
 extract_email_attachements(maildir = 'GSM_MTI', onlynew = TRUE)
 
}
}
