% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_BACKUP.R, R/db_backup.R
\name{mysqlrestore_host}
\alias{mysqlrestore_host}
\title{mysqlrestore_host}
\usage{
mysqlrestore_host(cnf = config::get(), backup, wipe = FALSE,
  restore_users = FALSE, parallel = TRUE, exclude)

mysqlrestore_host(cnf = config::get(), backup, wipe = FALSE,
  restore_users = FALSE, parallel = TRUE, exclude)
}
\arguments{
\item{cnf}{configuration variables are obtained from an external file config file. 
default to config::get().}

\item{backup}{path to backup dir. if missing the last backup is used.}

\item{wipe}{drop all non-system db-s before restore. default to FALSE}

\item{restore_users}{restore mysql.user table}

\item{parallel}{default to TRUE}

\item{exclude}{db to exclude from restoring}

\item{dir}{a directory containing all the sql files.}

\item{host}{default to localhost}

\item{user}{user}

\item{pwd}{pwd}

\item{...}{further options passed to mysqlrestore}

\item{wipe}{drop all non-system db-s before restore. default to FALSE}

\item{restore_users}{restore mysql.user table}

\item{parallel}{default to TRUE}

\item{...}{further options passed to mysqlrestore}
}
\description{
restore an entire db system or several db-s

restore an entire db system or several db-s
}
\examples{
\dontrun{
 fp = sdb::mysqldump_host('127.0.0.1',  'mihai', dir = '/home/mihai/Desktop')
 x = c('temp22')
 sdb::mysqlrestore_host(dir = fp ,user = 'mihai', wipe = TRUE, restore_users = TRUE, exclude = x)
}


\dontrun{

	require(dup)
 Sys.setenv(R_CONFIG_ACTIVE = "localhost")
 mysqlrestore_host()
}


}
