% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_through_email.R
\name{offlineimap}
\alias{offlineimap}
\title{offlineimap}
\usage{
offlineimap(cnf = config::get(), maildir)
}
\arguments{
\item{cnf}{configuration variables are obtained from an external file config file. 
default to config::get().}

\item{maildir}{maildir}
}
\description{
offlineimap
}
