% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_SNB.R
\name{read_boxtxt}
\alias{read_boxtxt}
\title{read SNB data}
\usage{
read_boxtxt(f)
}
\arguments{
\item{f}{path to the snb file}
}
\value{
a  \code{data.table} .
}
\description{
read an SNB file as data.table-s (only transponder and Light barrier data are parsed).
}
\examples{
x = read_boxtxt(system.file('test_files_SNB', '80', 'BOX0080.TXT', package = 'SNB2'))
}
\author{
MV
}
