% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUFFatSEEWIESEN.R
\name{RUFFatSEEWIESEN.photos_convert}
\alias{RUFFatSEEWIESEN.photos_convert}
\title{Convert, resize, and remove adult plumage photos.}
\usage{
RUFFatSEEWIESEN.photos_convert(ncores = 30)
}
\value{
number of converted files
}
\description{
Convert, resize, and remove adult plumage photos.
}
\seealso{
rw2webp
}
