% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rw2webp}
\alias{rw2webp}
\title{Convert RW2 photos to WebP}
\usage{
rw2webp(f, nam, width = 2000, crop = TRUE, cropstring = "80x80\%+0+0")
}
\arguments{
\item{nam}{output file name}

\item{width}{output width in px default to 2000}

\item{crop}{default to TRUE}

\item{cropstring}{magic geometry string, default to "80x80\%+0+0"}
}
\description{
Convert RW2 photos to WebP
}
