% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNB.R
\name{scidb_snbUpdater.b000}
\alias{scidb_snbUpdater.b000}
\title{\preformatted{     txt Box files (txt) to db b000 tables pipeline
}}
\usage{
scidb_snbUpdater.b000(cnf = config::get())
}
\arguments{
\item{cnf}{configuration variables are obtained from an external file config file.
default to config::get().}
}
\value{
\preformatted{   TRUE if all txt are updated
}
}
\description{
\preformatted{     txt Box files (txt) to db b000 tables pipeline
}
}
\examples{
\dontrun{
 require(dup)
 scidb_snbUpdater.b000 ()
}

}
