% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{scidb_snbUpdater.b000}
\alias{scidb_snbUpdater.b000}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{     txt Box files (txt) to db b000 tables pipeline
}\if{html}{\out{</div>}}}
\usage{
scidb_snbUpdater.b000(cnf = config::get())
}
\arguments{
\item{cnf}{configuration variables are obtained from an external file config file.
default to config::get().}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   TRUE if all txt are updated
}\if{html}{\out{</div>}}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     txt Box files (txt) to db b000 tables pipeline
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
 require(dup)
 scidb_snbUpdater.b000 ()
}

}
