% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try_outcome}
\alias{try_outcome}
\title{Try outcome
try_outcome}
\usage{
try_outcome(..., message)
}
\arguments{
\item{...}{one or several try() values}

\item{message}{to pass to push_msg}
}
\description{
Try outcome
try_outcome
}
