% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_backup_txt.R
\name{txtdump}
\alias{txtdump}
\title{txtdump}
\usage{
txtdump(db, table, remote = TRUE, dir = "~", cnf = config::get())
}
\arguments{
\item{db}{db name}

\item{table}{table name}

\item{remote}{when TRUE the file is uploaded to a remote host defined in cnf}

\item{dir}{directory path}

\item{cnf}{configuration variables (host, user, pwd, remotehost) are obtained 
from an external file config file. default to config::get().}
}
\value{
TRUE on success FALSE on failure
}
\description{
txtdump
}
\examples{
\dontrun{

	require(dup)
 Sys.setenv(R_CONFIG_ACTIVE = "localhost")
 txtdump(db = 'ARGOS', table = '2019_LBDO')
}

}
